created header files for non-exercise functions
