+++ synchronization_exercises/ejercicio_07/main.c
@@ -5,73 +5,12 @@
 #include <semaphore.h>
 #include <sys/types.h>
 #include <sys/syscall.h>
-#define LIMITE_LISTA 2
+#include "functions.h"
 #define INSTANCIAS_COMPILADOR 5
 
 sem_t semaforo_mutex_lista;
 sem_t semaforo_hay_resultados;
 sem_t semaforo_maximo_en_lista;
-
-typedef struct resultados
-{
-    int resultado_depositado;
-    int resultado;
-} resultados;
-
-resultados lista_resultados[LIMITE_LISTA];
-
-int obtener_nuevo_programa()
-{
-    printf("%d - El compilador esta obteniendo un nuevo programa...\n", (int)syscall(__NR_gettid));
-    return rand();
-}
-
-int compilar_programa(int id_programa)
-{
-    printf("%d - El compilador esta compilando el programa %d...\n", (int)syscall(__NR_gettid), id_programa);
-    return rand() % id_programa;
-}
-
-void depositar_resultado(int resultado, resultados lista_resultados[])
-{
-    int i = 0;
-    while (i < LIMITE_LISTA)
-    {
-        if (!lista_resultados[i].resultado_depositado)
-        {
-            lista_resultados[i].resultado_depositado = 1;
-            lista_resultados[i].resultado = resultado;
-
-            printf("%d - el compilador deposito el resultado %d en la lista de resultados...\n", (int)syscall(__NR_gettid), resultado);
-            sleep(1);
-            return;
-        }
-        i++;
-    }
-}
-
-int retirar_resultado(resultados lista_resultados[])
-{
-    int i = 0;
-    while (i < LIMITE_LISTA)
-    {
-        if (lista_resultados[i].resultado_depositado)
-        {
-            lista_resultados[i].resultado_depositado = 0;
-            printf("%d - el notificador retira el resultado %d...\n", (int)syscall(__NR_gettid), lista_resultados[i].resultado);
-            sleep(1);
-            return lista_resultados[i].resultado;
-        }
-        i++;
-    }
-}
-
-void enviar_email(int resultado)
-{
-    printf("%d - el notificador envia mail del resultado %d...\n", (int)syscall(__NR_gettid), resultado);
-}
-
-/* EJERCICIO */
 
 void *compilador()
 {
@@ -85,6 +24,8 @@ void *compilador()
         depositar_resultado(resultado, lista_resultados);
         sem_post(&semaforo_mutex_lista);
         sem_post(&semaforo_hay_resultados);
+
+        sleep(1);
     }
 }
 
@@ -99,6 +40,7 @@ void *notificador()
         sem_post(&semaforo_maximo_en_lista);
 
         enviar_email(resultado);
+        sleep(1);
     }
 }
